+++ source/projects/demo/demo.c
@@ -1,9 +1,14 @@
 #include "ext.h"
 #include "ext_obex.h"
 #include "cmx.h"
+#include "ext_dictionary.h"
+#include "ext_dictobj.h"
 
 typedef struct _demo {
     t_object ob;
+
+    void* outlet;         // outlet
+
     void *ob_proxy_1;     // inlet proxy
     void *ob_proxy_2;     // inlet proxy
     long ob_inletnum;     // # of inlet currently in use
@@ -26,7 +31,7 @@ void demo_bang(t_demo* x);
 void demo_float(t_demo *x, double f);
 t_max_err demo_name_get(t_demo *x, t_object *attr, long *argc, t_atom **argv);
 t_max_err demo_name_set(t_demo *x, t_object *attr, long argc, t_atom *argv);
-
+void demo_dict(t_demo* x, t_symbol* s);
 
 // global class pointer variable
 static t_class* demo_class = NULL;
@@ -40,6 +45,7 @@ void ext_main(void* r)
 
     class_addmethod(c, (method)demo_bang,  "bang", 0);
     class_addmethod(c, (method)demo_float, "float", A_FLOAT, 0);
+    class_addmethod(c, (method)demo_dict,  "dict",  A_SYM,   0);
 
     CLASS_ATTR_LABEL(c,     "name", 0,  "patch-wide name");
     CLASS_ATTR_SYM(c,       "name", 0,  t_demo, name);
@@ -57,7 +63,7 @@ void* demo_new(t_symbol* s, long argc, t_atom* argv)
     t_demo* x = (t_demo*)object_alloc(demo_class);
 
     if (x) {
-        outlet_new(x, NULL); // outlet
+        x->outlet = outlet_new(x, NULL); // outlet
 
         // inlet proxy
         x->ob_inletnum = 0;
@@ -131,10 +137,7 @@ void demo_bang(t_demo* x) {
     post("directory: %s", directory);
     post("filename: %s", filename);
 
-
-
-
-    outlet_bang(x->ob_proxy_1); 
+    outlet_bang(x->outlet); 
 }
 
 
@@ -154,6 +157,7 @@ t_max_err demo_name_get(t_demo *x, t_object *attr, long *argc, t_atom **argv)
     return MAX_ERR_NONE;
 }
 
+
 t_max_err demo_name_set(t_demo *x, t_object *attr, long argc, t_atom *argv)
 {
     if (argc && argv) {
@@ -162,4 +166,25 @@ t_max_err demo_name_set(t_demo *x, t_object *attr, long argc, t_atom *argv)
     }
     return MAX_ERR_NONE;
 }
+
+
+void demo_dict(t_demo* x, t_symbol* s)
+{
+    long        argc = 0;
+    t_atom      *argv = NULL;
+    t_max_err   err;
+
+    t_dictionary* d = dictionary_new();
+    err = dictionary_appendstring(d, s, "def");
+    err = dictobj_dictionarytoatoms(d, &argc, &argv);
+
+    if (argc && argv) {
+        // handles singles, lists, symbols, atomarrays, dictionaries, etc.
+        dictobj_outlet_atoms(x->outlet, argc, argv);
+    }
+
+    if (argv)
+        sysmem_freeptr(argv);
+}
+